Custom video player

+++ controllers/videoController.js
@@ -31,8 +31,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
+  req.user.videos.push(newVideo._id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 export const videoDetail = async (req, res) => {
@@ -40,7 +43,7 @@ export const videoDetail = async (req, res) => {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate('creator');
     res.render('videoDetail', { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -53,6 +56,10 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
+
+    if (video.creator !== req.user.id) {
+      throw new Error('Invaild User');
+    }
     res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
   } catch (error) {
     console.log(error);
@@ -76,6 +83,11 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
+    const video = await Video.findById(id);
+
+    if (video.creator !== req.user.id) {
+      throw new Error('Invaild User');
+    }
     await Video.findByIdAndRemove({ _id: id });
   } catch (error) {
     console.log(error);
